GPMixin: Fix handling of critical goals without function range
A function range does not make sense on a critical goal. Users were
forced to specify one anyway. This was due to

1. A check on whether a function range was specified; this check now
   properly skips critical goals.
2. RuntimeWarnings on comparisons with NaN for function range vs target
   checks. This section now properly skips critical goals.
3. A constraint of value NaN, because we mutiplied with the function
   range values still in __goal_hard_constraint(). We now multiply with 0.0
   in case of critical goals, forcing the target min/max to be the
   min/max constraints (barring any relaxation by the user).

Do note that critical goals still allow for "goal.relaxation" and
"constraint_relaxation" to be specified. It is also still allowed to
specify a function range for backwards compatibility, although it will
not actually be used. We might raise a warning about this in the future.

Also note that there was no actual testing done on critical goals.
Commit 07e4a342 did add a test model, but it did not actually add a test
case to run that model.