LinearizedOrderGPMixin: Use fast_newton to determine coefficients
Unlike IPOPT, the fast_newton approach is always available even if
CasADi was compiled with minimal external solver support.

Not a deciding factor for switching, but interesting nonetheless, is
that fast_newton takes about 444 us for an order-2 fit, whereas
nlpsol with IPOPT takes 16 ms.

Note that the resulting coefficients are not exactly the same.
Optimization results can therefore be slightly different.