OptimizationProblem: Fix block interpolation with interpolate()
Block interpolation only worked correctly for single values. The "mode"
argument was not properly passed along when interpolating for a
list/array of values, which is typically the case. It would instead to
linear interpolation (the default mode).

Note that internally we so far only use linear interpolation, which
worked as expected.

We add a bunch of tests to make sure that interpolation works as
intended.