OptimizationProblem: Optimize interpolate()
Since NumPy 1.16 the np.interp() method behaves the same way as our
internal __interpolate() implementation. The important bit for us is
that an exact match on the x value means that no interpolation is done,
and the corresponding y value is returned (whether it be np.inf, or
np.nan).

For linear interpolation, using np.interp() is always faster. There is
some additional overhead for making the call to NumPy, which seems to be
about 3 us from isolated timings. For a single value, that means that
the old implementation of __interpolate is about as fast as NumPy's
interp; for many values the latter is 10-100x faster, up until 400x
faster for equidistant data, or 3000x faster for non-equidistant data.

For block interpolation the NumPy interpolation using searchsorted() is
slower (7 us vs 3 us) for a single value. But with the manual
interpolation implementation scaling linearly, the searchsorted()
version is faster from 2 elements, up until 200x faster for many
elements.