Tests: Fix too strict equality test for linearized order goals
With the ordering of (floating point) operations not necessarily the
same between the two options in the test, the resulting answer can be
slightly different. It just so happened that until recently, the result
was exactly equal. We now check for approximate equality instead, which
is what we should have been doing in the first place.