Add test for the examples; CI runs for tags
We add a test that searches the example folder for files in src
subfolders containing 'example' in their filename. These are run and
this test fails when any errors/exceptions occur.

This test makes sure that all examples can be run. It does _not_ test
the output of the examples, as many examples have no unique solution.

Note that we do not run these tests by default, but only for tags (or
when invoked locally by the user by e.g. `tox -e examples37`). The
reason is that it takes about ~100 seconds to run the examples, compared
to 30 seconds for all (other) unittests.