CollInt: Fix bug when history values differ in length
The flattening of the history values list of arrays failed when those
arrays have different lengths. Instead of relying on np.unique's
automatic flattening, we just concatenate them ourselves. Note that when
concatenating, we have to be careful to support the case where there is
no history at all. We do this by including an empty array in the
concatenation.