CollInt: Simplify discretize_states() and discretize_controls()
Instead of recalculating the correct state vector offset over and over
again, we now reuse the indices we calculated at the start of
discretize_states() or discretize_controls() as much as possible in
these methods and their corresponding extract_* methods. This reduces
the chance for mistakes in offset calculations, and also opens up the
possibility to more easily change the location of a variable's state
vector entries.