CollInt: Simplify and optimize state_at() and control_at()
Instead of recalculating the offsets and looping until we find a
matching variable, we can simply reuse the indices for a variable we
already stored.

Because state_at() also works for controls, we can now just forward a
call to control_at() to state_at(). Note that we do check if it actually
is a control, and raise a KeyError otherwise, for backwards compability
(and to avoid confusion).

By relying on the indices, we can also get rid of control_at() in
ControlTreeMixin, as the base implementation in CollInt now works just
as well.

Closes #1067