CollInt: Simplify nominals for initial derivatives
Instead of recalculating the nominal over and over again based on the
initial time step size, we can just calculate the initial derivative
nominals once. This makes handling elsewhere in the code much more
similar to that of "regular" variables.

Note that this commit also "fixed" ModelicaMixin to properly call
super() for variable_nominal().