Storage: Allow overwriting reference datetime with itself when fixed
We fix the reference datetime as soon as get/set_timeseries_sec is
called. We did not allow any further setting attemps on it, even when
the value would end up not being changed.

When calling optimize() multiple times, e.g. for initial state
estimation, we could end up setting the reference datetime _after_ it
has already been fixed by calling pre() multiple times. Working around
this exception is very difficult as things are now. It is much easier to
allow setting the reference datetime after it is fixed when the result
means nothing will change.