GPMixin: Fix goal is_empty check
If one of the targets is np.nan, and the other a Timeseries containing
all np.nan (inactive), we would get an Exception thrown due to a size
error. This is because this type of goal, although empty/inactive, was
not being skipped in processing.

This commit fixes the goal.is_empty check, properly skipping such goals.
We also add tests for the expected behavior of is_empty, e.g. that it
still returns False when dealing with minimization goals (both targets
np.nan).

Closes #1098