SinglePassGPMixin: Fix priority of applied constraint_relaxation
The behavior in GoalProgrammingMixin with keep_soft_constraints=True is
that the constraint_relaxation of priority 2 is applied to the objective
of priority 2. This is because we call `goal_programming_options` in
GoalProgrammingMixin at the end of priority 2 to make this constraint.

In SinglePassGoalProgrammingMixin, we would get the
constraint_relaxation by calling `goal_programming_options` at the start
of the next priority, so the constraint_relaxation of priority 3 would
be applied to the objective of priority 2. This is not necessarily
wrong, but does differ from GoalProgrammingMixin. We therefore fix the
discrepancy by making SinglePassGoalProgrammingMixin behave equal in
this sense to GoalProgrammingMixin; the constraint relaxation at
priority 2 is applied to the objective of priority 2.