Tests/CSVLookupTableMixin: Do not generate plots
The generation of plots requires matplotlib, which is not an explicit
dependency of our test environment. Instead of adding matplotlib to our
test dependencies, we can just disable the generation of plots, as it
does not add anything to the tests.

Note that the generation of plots might make sense when debugging why
the test is failing, but that is a different use case for which the user
can easily just install matplotlib.