HomotopyMixin: Fix overwriting of seed
It is necessary to overwrite the seed within the HomotopyMixin class
whenever the results of the latest optimization run are stored here.
Note that a necessary requirement for this is that theta must be greater
than zero. Moreover, there are only two situations in which we need to
overwrite the seed: when the class GoalProgrammingMixin is not used, or
whenever it is the first run of the goal programming loop. To detect
this, and to make it clear that we internally rely on this variable, we
rename __first_run to _gp_first_run.