Add flag to run optional debug checks (private API)
When setting up or debugging a model it can be useful to do typical
checks to see if you made a mistake. These checks can be rather
computationally expensive, which is why it should be easy to disable
them when running operationally. A (for now private) flag
`_debug_check_level` is added to OptimizationProblem/SimulationProblem
that can be used to enable/disable certain checks.

Foreseeing quite a few of these checks, it should be easy to filter
checks based on verbosity and how expensive they are to perform. A
typical way of doing this is with an enum, just like logging levels
(DEBUG, INFO, WARNING, etc). To this end the value of
`_debug_check_level` can be set to the value of a `DebugLevel` enum.
This enum has with 5 levels, with room for future refinement. The
default check level is in the middle (MEDIUM), with two available below
(NONE, LOW) and two above (HIGH, VERYHIGH).

With more of these checks, and considering how expensive and useful they
might be for a certain type of issue, it is also possible to filter more
precise by setting the value of `_debug_check_level` to a function. That
way one can filter a specific test on its level and fully qualified
name.

For some checks it might make sense to accept some arguments. For
example, a certain tolerance value above which messages should be
logged. For this purpose the `_debug_check_options` class variable of
OptimizationProblem/SimulationProblem can be used. This is a dictionary
that maps a fully specified check name to (a dictionary of) additional
keyword arguments.

Note that it is likely that some details of these debug checks will
change, which is why it is only part of the private API for now.

Closes #1042