Add debug check on jacobian coefficients at x0
Many solvers (IPOPT, CPLEX, etc) would prefer the coefficients in the
Jacobian to be in the order of [1E-2, 1E2], and also for the range of
coefficients for a certain variable or in a certain constraint to be
rather limited (say 1E3). The check added in this commit gives feedback
to the user on what constraints/variables do not meet these targets.

Note that not satisfying these limits/ranges does not mean that the
problem will not solve or even be slower, because solvers tend to be
quite smart in how they themselves rescale equations or evade numerical
instabilities. That said, the larger (more variables, more constraints)
a problem gets, the more sensitive solvers tend to become to bad
scaling.