Add debug check logging whether constraints are linear/quadratic/other
It is often useful to make sure that the constraints are linear, such
that convexity is guaranteed. Depending in the options of for example
GoalProgrammingMixin, it can be possible that a quadratic objective in
one priority leads to a quadratic constraint in the next priority, which
a solver may not be able to solve (or may not even be convex anymore).

Eventually this check might be rewritten such that it (also) logs
whether a problem is (MI)LP/QP/QCQP/NLP, and also say something about
the convexity.