NetCDF: Fix conversion of imported times to datetimes
`NetCDF4` depends on `cftime`, the latter released a version 1.1 which
changed the behavior of the `num2date` function. It also added a new
convenience method `num2pydate` which seems to sometimes lose the
millisecond information.

We only implicitly depend on `cftime` via `NetCDF4`, but with this
breaking change on their end we need to make sure that the correct
version is installed. We therefore add `cftime` to our list of
(optional) dependencies, and rely on the version we know is correct
(which excludes 1.1.0).

See e.g. https://github.com/Unidata/cftime/issues/134 for details on why
this breaking change was made.

See https://github.com/Unidata/cftime/issues/144 for the issue on
num2pydate losing millisecond information.