Storage: Optimize calls to .times_sec property
Calling `self.io.times_sec` would make a copy, to be sure that the user
could not overwrite the internal array of times. Especially in
simulation, where we can do many time steps, this copy() call was rather
expensive. The same intent can be accomplished by disallowing writes to
the array we return.

Note that the times in seconds are immutable as soon as they are
generated (which is usually done when a request is done like
`get_timeseries_sec`). At that point, they are no longer allowed to
change.