Simulation: Optimize update/rootfinder call by avoiding vertcat
Vertcat calls are relatively expensive to do. The vertcat call the
prepends `dt` was therefore taking a substantial amount of time, equal
or more than the actual solving of the equations with rootfinder for
linear models.

Instead of vertcatting, we can just specify the `dt` as additional
parameter to the ca.Function that we pass to rootfinder. We can then
also just pass it separately when calling `__do_step` to do a rootfinder
step.