Style fix for LookupTable class
A method in OptimizationProblem was returning a "LookupTable" type, with
the class not defined anywhere in scope. One of the CI/CD style checkers
became a bit more aggressive recently and started to catch this as an
error.

We now properly defined a base LookupTable class, from which the one in
csv_lookup_table_mixin.py inherits.