OptimizationProblem/ModelicaMixin: Support string parameters/constants
Because string parameters and constants are not part of the DAE, they
are separate properties of a pymoca Model. This separation is kept in
RTC-Tools, to make a clear distinction between numeric parameters that
may or may not be part of the DAE or other constraints, and string
parameters that definitely are not.

Note that in RTC-Tools we do not make a distiction between string
constants and parameters, so we cluster both together.