Tests: Fix definition of delayed symbol as non-control input
The definition of `x_delayed` should be the same as that of `x`. In
ensemble mode, control inputs are allocated to the same state vector
entries, whereas states/alg_states are not. That means that for the
TestCSVMixinEnsemble test, both `u` and `x_delayed` would have shared
indices/symbols for each time step between the ensemble members, but `x`
would not. This is obviously wrong, as `x_delayed` should also have
separate symbols per ensemble member just like `x`. This incongruency
easily led to an overdetermined system of equations. This commit fixes
the issue by removing the `input` + `fixed=False` combination that
(wrongly) marked `x_delayed` as a control input.

Note that in ModelWithInitial and ModelWithInitialLinear the definition of
`x_delayed` was already correct (as a non-control variable).

Also note that we turn it into an output, such that the NetCDF tests
still write it out.