HomotopyMixin: Add support for specifying starting theta
Sometimes it can be useful to skip the `theta = 0`, e.g. when a good
seed is available from a previous run. This commit adds an option
`theta_start` for the user to specify what theta to start the homotopy
process at. The default value of this new option is the current (before
this commit) starting theta value of 0.