CollInt: Fix warning for delay expressions having no history
In case partial history was present, the delay expressions with missing
values were correctly logged a warning for. However, in case no history
was provided at all, these warnings were not logged due to the if-check
not catching this. This commit makes sure that warnings are logged also
in that scenario.

Closes #1130