Simulation: Make extract_results() return a dict of np.arrays
For performance reasons, the internal storage was changed from NumPy
arrays to lists in 516cd10f. The extract_results() method then also
started returning lists, which is different from the equivalent method
in optimization that returns arrays. This is somewhat confusing, and can
trip up code that is shared between simulation and optimization
implementations.

This commit brings simulation's extract_results() in line with that of
otpimization by returning NumPy arrays as well. Note that we keep the
internal storage as lists for performance reasons. As such, a call to
extract_results() is expensive, and should be done as little as
possible, e.g. once in post().