ControlTree: Add support for k-ary trees with k >= 10
A string representation of the branch indices was used, and subsequently
the _length_ of said string was used in the splitting logic as well.
This works fine for k <= 9, but no longer matches the number of splits
(or uniquely identifies the branch) for k >= 10.

This commit turns the concatenated string into a tuple, such that the
logic holds for k of arbitrary magnitude.