optimization_problem: option error as info
We add the option to log an error message after the solver does not
succeed as an info message rather than an error. This should be used
 for example if FEWS searches for error occurences and you want
  these to be ignored.
How to use:
In the pre() of you model define
        self.loop_over_error = []
        self.loop_over_error.append([priority, True/False, 'error'])
where priority is the priority for which you want an info message
 instead of an error. True if you want to edit the logging, false if not.
 in error you can specify the error message you would like to ignore.
For example 'maxnodes' if you want to recieve the solver timout as
an info message. The default is such that if no information is
provided for the error and second option is set to true then all
types of solver return status will be an info message. if no such list
exists for a certain priority then solver succeeds will be info and if
not an error message.