test: add unit test for IOMixin.times
Add a unit test for the IOMixin times method.
This also tests that the correct number of times
is returned when optimizing for only a subset of timesteps.
In particular, it tests that only the first time is returned
when optimizing for zero time steps.
This is useful when checking the initial conditions.