style/flake8: Make default flake8 style exceptions explicit
By default flake8 ignores these warnings. Because this list can change
from version to version, and because we might want to add our own
ignores, we make the current default list explicit.

Note that some of these ignores are not needed, and we will remove them
in a later commit.