style/flake8: Reduce max line length to 100
Preparing for black, which in its documentation states that "In general,
90-ish seems like the wise choice" and "people with sight disabilities
find it harder to work with line lengths exceeding 100 characters". By
"90-ish" it's meant that about 10% longer is allowed. Those two together
makes that we reduce the maximum line length from 120 to 100.

Note that we add an ignore for max line length, as we will rely on black
to fix most of the offending lines.