style: Prepare for black, keep arguments over multiple lines
To prevent black from putting these arguments on a single line, we put a
comma at the end of the last argument/class.

Note that there are few locations in the code where we have split
arguments over multiple lines like this, let alone where black
"disagrees" and will put them on a single line. The change in this
commit is basically the only location where that is the case.