SimulationProblem: Support models with empty DAE
If we have a model in which the only equation is an aliasing equation,
or a model in which we will define most equations in Python, the DAE
will be empty.

Note that the code already support the `__initial_residual` to be empty
(e.g. when there were no "der(...)" statements in the model), but not
yet when there were no equations at all.