Style: Switch from flake8 to ruff
With the move to pre-commit, we want the style checks to be fast, and we
also want to be able to get them automatically fixed where possible. We
therefore move from Flake8 to Ruff, while trying to keep most of the
checks the same. The only really incompatible change is from
flake8-import-order to Ruff's isort; and here the former's behavior was
the outlier compared to the more popular isort.