Style: Import order/isort fixes by Ruff
These are the changes Ruff makes to make the code style compliant to its
defaults, compared to the defaults of our flake8 configuration. Mostly
it affects the import order, in which flake8-import-order also differed
from isort in that it separated different third party imports (e.g.
casadi and numpy) with a blank line.