examples: remove output from version control
We do not want example output in version control,
because the output often changes due to small changes in the solver.

Therefore, this commit moves the contents to a folder called
`reference_folder`,
as we still need the data for plots in the documentation.

Doing so also keeps the option open of reference checks/tests
on the examples in the future.

Note that the `output` folders themselves are not removed
(kept around with a `.gitignore` in them),
because running an example will require these folders to exist.