workflows: add work flow based on the gitlab ci workflow
Add a github workflow based on the gitlab CI workflow.

Private tokens to access codecov and pypi are still missing.

Tests are no longer executed with python 3.8,
since this version is almost end-of-life
and seems to sometimes result in a mysterious crash
during package installation, resulting in a message
"THESE PACKAGES DO NOT MATCH THE HASHES FROM THE REQUIREMENTS FILE"
for an "unknown package".