Remove dependency on deprecated pkg_resources
The pkg_resources package has been removed from the Python standard
library in version 3.12, while still available when (manually)
installing setuptools. A more modern approach is the usage of
"importlib", available since Python 3.9. Note that we do require a
change to "entry_point" accepting a group parameter that was introduced
with Python 3.10, so we fall back to a version of "importlib_metadata"
that support this. Also note that we also only require this external
(backport) dependency for Python 3.9.