sim: Do not warn out-of-bounds start value if not explicitly set
We would warn users about initial/start values they never explicitly
set. The log message is specifically about states that are minimized
(not constrained/bounded), so 'out-of-bounds' start values would not
lead to infeasible problems. To not needlessly confuse the user with
warnings, we change the message to "DEBUG" level, unless they explicitly
set a start value themselves in which case we keep the "WARNING" level.

Note that there is no good way (yet) to separate a default value of 0
coming from Pymoca from an explicitly specified value of 0. The
assumption (unchanged in this commit) is that 0 means 'default'.