CollInt: Support returning times from 'states_in()'
Internally the `times` were already returned with the `__states_time_in`
method. Exposing this functionality to the user will allow them to e.g.
compute integrals/sums of algebraic states/control inputs. We do this by
adding an extra argument (which _has_ to be passed as keyword), default
False to match the old behavior.

Our implementation of `integral` takes averages over a time step, which
makes sense for state variables like reservoir volumes. But if one is
interested in a sum/average of algebraic states, they would want to
exclude the value at t0 (assuming backwards euler). This is non-trivial,
and would mean the user has to reimplent the logic in `__states_time_in`
for history etc.

Note that we may want to support different ways of calculating the
integral directly in `integral()` in the future, but at least optionally
returning the times from `states_in()` in this commit already allows
this and more in a generic way.