CollInt/casadi_helpers: Fix interpolation for single-timestep variables
Other interpolation tests already cover a variable with different time
steps than the overall collocation time steps. In the case where this
'different' means just a _single_ time step, the interpolation with
CasADi would fail.

Note that 'interp1d' already extrapolated outside the range of `ts`, it
just refused to do so for a single timestep.