casadi_helpers/substitute_in_external: Robustify resolving to numeric values
Complicated symbolic MX expressions including function calls do not
resolve well to something that can be checked for 'is_constant()' (e.g.
in ModelicaMixin).

Expanding the function helps with this, at a slight overhead. We also
have to be sure to loop when resolving the value. For example, a
symbolic bound of 'u_min' might in turn evaluate to a symbolic
expression 'a * b'.  We keep going until we have a numeric value, up
until a depth of MAX_SUBSTITUTE_DEPTH (=10).